# Vendor tsdb and reuse iterators