--- a/pkg/ingester/ingester.go
+++ b/pkg/ingester/ingester.go
@@ -20,6 +20,7 @@ import (
 	"github.com/prometheus/prometheus/pkg/labels"
 
 	cortex_chunk "github.com/cortexproject/cortex/pkg/chunk"
+	"github.com/cortexproject/cortex/pkg/chunk/encoding"
 	"github.com/cortexproject/cortex/pkg/ingester/client"
 	"github.com/cortexproject/cortex/pkg/ring"
 	"github.com/cortexproject/cortex/pkg/util"
@@ -254,11 +255,11 @@ func (i *Ingester) StopIncomingRequests() {
 
 // Push implements client.IngesterServer
 func (i *Ingester) Push(ctx old_ctx.Context, req *client.WriteRequest) (*client.WriteResponse, error) {
-	var lastPartialErr error
-
+	var lastPartialErr, err error
+	var reuseIter encoding.Iterator
 	for _, ts := range req.Timeseries {
 		for _, s := range ts.Samples {
-			err := i.append(ctx, ts.Labels, model.Time(s.TimestampMs), model.SampleValue(s.Value), req.Source)
+			reuseIter, err = i.append(ctx, ts.Labels, model.Time(s.TimestampMs), model.SampleValue(s.Value), req.Source, reuseIter)
 			if err == nil {
 				continue
 			}
@@ -279,39 +280,41 @@ func (i *Ingester) Push(ctx old_ctx.Context, req *client.WriteRequest) (*client.
 	return &client.WriteResponse{}, lastPartialErr
 }
 
-func (i *Ingester) append(ctx context.Context, labels labelPairs, timestamp model.Time, value model.SampleValue, source client.WriteRequest_SourceEnum) error {
+func (i *Ingester) append(ctx context.Context, labels labelPairs, timestamp model.Time,
+	value model.SampleValue, source client.WriteRequest_SourceEnum, reuseIter encoding.Iterator) (encoding.Iterator, error) {
 	labels.removeBlanks()
 
 	i.stopLock.RLock()
 	defer i.stopLock.RUnlock()
 	if i.stopped {
-		return fmt.Errorf("ingester stopping")
+		return reuseIter, fmt.Errorf("ingester stopping")
 	}
 
 	i.userStatesMtx.RLock()
 	defer i.userStatesMtx.RUnlock()
 	state, fp, series, err := i.userStates.getOrCreateSeries(ctx, labels)
 	if err != nil {
-		return err
+		return reuseIter, err
 	}
 	defer func() {
 		state.fpLocker.Unlock(fp)
 	}()
 
 	prevNumChunks := len(series.chunkDescs)
-	if err := series.add(model.SamplePair{
+	reuseIter, err = series.add(model.SamplePair{
 		Value:     value,
 		Timestamp: timestamp,
-	}); err != nil {
+	}, reuseIter)
+	if err != nil {
 		if mse, ok := err.(*memorySeriesError); ok {
 			state.discardedSamples.WithLabelValues(mse.errorType).Inc()
 			if mse.noReport {
-				return nil
+				return reuseIter, nil
 			}
 			// Use a dumb string template to avoid the message being parsed as a template
 			err = httpgrpc.Errorf(http.StatusBadRequest, "%s", mse.message)
 		}
-		return err
+		return reuseIter, err
 	}
 
 	memoryChunks.Add(float64(len(series.chunkDescs) - prevNumChunks))
@@ -325,7 +328,7 @@ func (i *Ingester) append(ctx context.Context, labels labelPairs, timestamp mode
 		state.ingestedAPISamples.inc()
 	}
 
-	return err
+	return reuseIter, err
 }
 
 // Query implements service.IngesterServer