--- a/pkg/querier/batch/chunk.go
+++ b/pkg/querier/batch/chunk.go
@@ -16,7 +16,7 @@ type chunkIterator struct {
 
 func (i *chunkIterator) reset(chunk chunk.Chunk) {
 	i.chunk = chunk
-	i.it = chunk.Data.NewIterator()
+	i.it = chunk.Data.NewIterator(i.it)
 	i.batch.Length = 0
 	i.batch.Index = 0
 }