This includes https://github.com/prometheus/tsdb/pull/642. I have added similar re-use of iterators here. There is a possibility of allocs improvements in ingester push and I am trying to realize that.

I will post the benchmark results soon.