@codesome 2019-08-12
This is the benchmark results of existing benchmark. Mostly better for the `BigChunk`. Also the gains are more when the number of chunks the query has to parse is more (see the next benchmark).
```
benchmark                                                                     old ns/op     new ns/op     delta
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/DoubleDelta/matrixes/-8       124279606     126273868     +1.60%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/DoubleDelta/iterators/-8      137742086     137891187     +0.11%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/DoubleDelta/batches/-8        115307841     112422090     -2.50%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Varbit/matrixes/-8            124573115     126482215     +1.53%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Varbit/iterators/-8           137869281     135199069     -1.94%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Varbit/batches/-8             118543740     114143517     -3.71%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Bigchunk/matrixes/-8          143376091     138695368     -3.26%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Bigchunk/iterators/-8         154839197     149437380     -3.49%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Bigchunk/batches/-8           134870811     130754059     -3.05%
BenchmarkChunkQueryable/foo/step_1m0s/DoubleDelta/matrixes/-8                 122262673     121712943     -0.45%
BenchmarkChunkQueryable/foo/step_1m0s/DoubleDelta/iterators/-8                135786304     131509847     -3.15%
BenchmarkChunkQueryable/foo/step_1m0s/DoubleDelta/batches/-8                  113268287     110935556     -2.06%
BenchmarkChunkQueryable/foo/step_1m0s/Varbit/matrixes/-8                      121139305     118175574     -2.45%
BenchmarkChunkQueryable/foo/step_1m0s/Varbit/iterators/-8                     133062550     129900050     -2.38%
BenchmarkChunkQueryable/foo/step_1m0s/Varbit/batches/-8                       114257016     110847201     -2.98%
BenchmarkChunkQueryable/foo/step_1m0s/Bigchunk/matrixes/-8                    136413422     134883931     -1.12%
BenchmarkChunkQueryable/foo/step_1m0s/Bigchunk/iterators/-8                   152123490     147195249     -3.24%
BenchmarkChunkQueryable/foo/step_1m0s/Bigchunk/batches/-8                     129592206     126676483     -2.25%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/DoubleDelta/matrixes/-8      33166641      33091968      -0.23%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/DoubleDelta/iterators/-8     16549587      16392219      -0.95%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/DoubleDelta/batches/-8       21650928      20726741      -4.27%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Varbit/matrixes/-8           31798914      33051433      +3.94%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Varbit/iterators/-8          19659756      19404448      -1.30%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Varbit/batches/-8            23650084      22841007      -3.42%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Bigchunk/matrixes/-8         48951225      48307891      -1.31%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Bigchunk/iterators/-8        34573840      33511635      -3.07%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Bigchunk/batches/-8          39301950      37875161      -3.63%
BenchmarkChunkQueryable/foo/step_10m0s/DoubleDelta/matrixes/-8                33992465      33830655      -0.48%
BenchmarkChunkQueryable/foo/step_10m0s/DoubleDelta/iterators/-8               31024230      30431688      -1.91%
BenchmarkChunkQueryable/foo/step_10m0s/DoubleDelta/batches/-8                 32957156      31802592      -3.50%
BenchmarkChunkQueryable/foo/step_10m0s/Varbit/matrixes/-8                     32607443      33417103      +2.48%
BenchmarkChunkQueryable/foo/step_10m0s/Varbit/iterators/-8                    31519665      31012217      -1.61%
BenchmarkChunkQueryable/foo/step_10m0s/Varbit/batches/-8                      32630591      31587815      -3.20%
BenchmarkChunkQueryable/foo/step_10m0s/Bigchunk/matrixes/-8                   50112043      50352803      +0.48%
BenchmarkChunkQueryable/foo/step_10m0s/Bigchunk/iterators/-8                  48401702      46832878      -3.24%
BenchmarkChunkQueryable/foo/step_10m0s/Bigchunk/batches/-8                    60351191      59115913      -2.05%

benchmark                                                                     old allocs     new allocs     delta
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/DoubleDelta/matrixes/-8       528656         528659         +0.00%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/DoubleDelta/iterators/-8      520784         520785         +0.00%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/DoubleDelta/batches/-8        520034         518603         -0.28%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Varbit/matrixes/-8            527937         527937         +0.00%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Varbit/iterators/-8           520065         520063         -0.00%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Varbit/batches/-8             519314         518598         -0.14%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Bigchunk/matrixes/-8          532971         529377         -0.67%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Bigchunk/iterators/-8         524383         520787         -0.69%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Bigchunk/batches/-8           523632         520038         -0.69%
BenchmarkChunkQueryable/foo/step_1m0s/DoubleDelta/matrixes/-8                 528626         528626         +0.00%
BenchmarkChunkQueryable/foo/step_1m0s/DoubleDelta/iterators/-8                520755         520755         +0.00%
BenchmarkChunkQueryable/foo/step_1m0s/DoubleDelta/batches/-8                  520005         518573         -0.28%
BenchmarkChunkQueryable/foo/step_1m0s/Varbit/matrixes/-8                      527906         527905         -0.00%
BenchmarkChunkQueryable/foo/step_1m0s/Varbit/iterators/-8                     520035         520036         +0.00%
BenchmarkChunkQueryable/foo/step_1m0s/Varbit/batches/-8                       519287         518570         -0.14%
BenchmarkChunkQueryable/foo/step_1m0s/Bigchunk/matrixes/-8                    532941         529345         -0.67%
BenchmarkChunkQueryable/foo/step_1m0s/Bigchunk/iterators/-8                   524356         520758         -0.69%
BenchmarkChunkQueryable/foo/step_1m0s/Bigchunk/batches/-8                     523604         520009         -0.69%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/DoubleDelta/matrixes/-8      62092          62092          +0.00%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/DoubleDelta/iterators/-8     54220          54220          +0.00%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/DoubleDelta/batches/-8       53470          52038          -2.68%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Varbit/matrixes/-8           61372          61371          -0.00%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Varbit/iterators/-8          53500          53500          +0.00%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Varbit/batches/-8            52750          52034          -1.36%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Bigchunk/matrixes/-8         66407          62811          -5.42%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Bigchunk/iterators/-8        62151          58553          -5.79%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Bigchunk/batches/-8          61401          57804          -5.86%
BenchmarkChunkQueryable/foo/step_10m0s/DoubleDelta/matrixes/-8                62064          62064          +0.00%
BenchmarkChunkQueryable/foo/step_10m0s/DoubleDelta/iterators/-8               54193          54193          +0.00%
BenchmarkChunkQueryable/foo/step_10m0s/DoubleDelta/batches/-8                 53443          52011          -2.68%
BenchmarkChunkQueryable/foo/step_10m0s/Varbit/matrixes/-8                     61344          61344          +0.00%
BenchmarkChunkQueryable/foo/step_10m0s/Varbit/iterators/-8                    53473          53473          +0.00%
BenchmarkChunkQueryable/foo/step_10m0s/Varbit/batches/-8                      52723          52007          -1.36%
BenchmarkChunkQueryable/foo/step_10m0s/Bigchunk/matrixes/-8                   66380          62784          -5.42%
BenchmarkChunkQueryable/foo/step_10m0s/Bigchunk/iterators/-8                  62125          58528          -5.79%
BenchmarkChunkQueryable/foo/step_10m0s/Bigchunk/batches/-8                    65444          57779          -11.71%

benchmark                                                                     old bytes     new bytes     delta
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/DoubleDelta/matrixes/-8       84122153      84122370      +0.00%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/DoubleDelta/iterators/-8      21579672      21579760      +0.00%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/DoubleDelta/batches/-8        21678304      21586768      -0.42%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Varbit/matrixes/-8            84122171      84122350      +0.00%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Varbit/iterators/-8           21579801      21579445      -0.00%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Varbit/batches/-8             21678343      21586602      -0.42%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Bigchunk/matrixes/-8          84536344      84191185      -0.41%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Bigchunk/iterators/-8         21925086      21579972      -1.57%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Bigchunk/batches/-8           22023737      21678607      -1.57%
BenchmarkChunkQueryable/foo/step_1m0s/DoubleDelta/matrixes/-8                 84081008      84081044      +0.00%
BenchmarkChunkQueryable/foo/step_1m0s/DoubleDelta/iterators/-8                21539208      21539196      -0.00%
BenchmarkChunkQueryable/foo/step_1m0s/DoubleDelta/batches/-8                  21637897      21546283      -0.42%
BenchmarkChunkQueryable/foo/step_1m0s/Varbit/matrixes/-8                      84080985      84080964      -0.00%
BenchmarkChunkQueryable/foo/step_1m0s/Varbit/iterators/-8                     21539244      21539259      +0.00%
BenchmarkChunkQueryable/foo/step_1m0s/Varbit/batches/-8                       21638000      21546356      -0.42%
BenchmarkChunkQueryable/foo/step_1m0s/Bigchunk/matrixes/-8                    84495275      84150040      -0.41%
BenchmarkChunkQueryable/foo/step_1m0s/Bigchunk/iterators/-8                   21884849      21539550      -1.58%
BenchmarkChunkQueryable/foo/step_1m0s/Bigchunk/batches/-8                     21983456      21638276      -1.57%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/DoubleDelta/matrixes/-8      65458109      65458061      -0.00%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/DoubleDelta/iterators/-8     2915333       2913837       -0.05%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/DoubleDelta/batches/-8       3013739       2921682       -3.05%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Varbit/matrixes/-8           65458028      65457884      -0.00%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Varbit/iterators/-8          2913831       2912745       -0.04%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Varbit/batches/-8            3013763       2922055       -3.04%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Bigchunk/matrixes/-8         65872172      65527126      -0.52%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Bigchunk/iterators/-8        3670991       3322463       -9.49%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Bigchunk/batches/-8          3770699       3424253       -9.19%
BenchmarkChunkQueryable/foo/step_10m0s/DoubleDelta/matrixes/-8                65417942      65417936      -0.00%
BenchmarkChunkQueryable/foo/step_10m0s/DoubleDelta/iterators/-8               2875991       2875970       -0.00%
BenchmarkChunkQueryable/foo/step_10m0s/DoubleDelta/batches/-8                 2974690       2883004       -3.08%
BenchmarkChunkQueryable/foo/step_10m0s/Varbit/matrixes/-8                     65417928      65417921      -0.00%
BenchmarkChunkQueryable/foo/step_10m0s/Varbit/iterators/-8                    2875959       2875978       +0.00%
BenchmarkChunkQueryable/foo/step_10m0s/Varbit/batches/-8                      2974725       2883040       -3.08%
BenchmarkChunkQueryable/foo/step_10m0s/Bigchunk/matrixes/-8                   65832258      65487068      -0.52%
BenchmarkChunkQueryable/foo/step_10m0s/Bigchunk/iterators/-8                  3637383       3292071       -9.49%
BenchmarkChunkQueryable/foo/step_10m0s/Bigchunk/batches/-8                    4126833       3390834       -17.83%
```

And this is for a 650MB tarball chunk for a query as described in this comment https://github.com/prometheus/prometheus/pull/5707#issuecomment-505440230
```
benchmark                                                                                                                                                        old ns/op       new ns/op       delta
BenchmarkChunkQueryableFromTar/query=count(http_server_requests_seconds_bucket{origin_prometheus="devstage"}),from=1561346880,to=1561368480,step=15.000000-8     39286990093     37987331824     -3.31%

benchmark                                                                                                                                                        old allocs     new allocs     delta
BenchmarkChunkQueryableFromTar/query=count(http_server_requests_seconds_bucket{origin_prometheus="devstage"}),from=1561346880,to=1561368480,step=15.000000-8     6481938        2860277        -55.87%

benchmark                                                                                                                                                        old bytes      new bytes      delta
BenchmarkChunkQueryableFromTar/query=count(http_server_requests_seconds_bucket{origin_prometheus="devstage"}),from=1561346880,to=1561368480,step=15.000000-8     5516013432     5168349984     -6.30%
```