@codesome 2019-08-14
Updated numbers for BigChunk after last few changes
```
benchmark                                                                  old ns/op     new ns/op     delta
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Bigchunk/matrixes/-8       139935257     137669154     -1.62%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Bigchunk/iterators/-8      152545075     151154759     -0.91%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Bigchunk/batches/-8        134357445     131323914     -2.26%
BenchmarkChunkQueryable/foo/step_1m0s/Bigchunk/matrixes/-8                 137261435     135783993     -1.08%
BenchmarkChunkQueryable/foo/step_1m0s/Bigchunk/iterators/-8                149342408     148639332     -0.47%
BenchmarkChunkQueryable/foo/step_1m0s/Bigchunk/batches/-8                  128164516     130506813     +1.83%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Bigchunk/matrixes/-8      48899817      48609061      -0.59%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Bigchunk/iterators/-8     34260214      33673446      -1.71%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Bigchunk/batches/-8       38609613      38634906      +0.07%
BenchmarkChunkQueryable/foo/step_10m0s/Bigchunk/matrixes/-8                50076761      49687183      -0.78%
BenchmarkChunkQueryable/foo/step_10m0s/Bigchunk/iterators/-8               48340936      48039758      -0.62%
BenchmarkChunkQueryable/foo/step_10m0s/Bigchunk/batches/-8                 60095097      57575945      -4.19%

benchmark                                                                  old allocs     new allocs     delta
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Bigchunk/matrixes/-8       532972         529376         -0.67%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Bigchunk/iterators/-8      524383         520788         -0.69%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Bigchunk/batches/-8        523634         519322         -0.82%
BenchmarkChunkQueryable/foo/step_1m0s/Bigchunk/matrixes/-8                 532941         529346         -0.67%
BenchmarkChunkQueryable/foo/step_1m0s/Bigchunk/iterators/-8                524354         520760         -0.69%
BenchmarkChunkQueryable/foo/step_1m0s/Bigchunk/batches/-8                  523605         519294         -0.82%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Bigchunk/matrixes/-8      66407          62812          -5.41%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Bigchunk/iterators/-8     62151          58553          -5.79%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Bigchunk/batches/-8       61400          57088          -7.02%
BenchmarkChunkQueryable/foo/step_10m0s/Bigchunk/matrixes/-8                66380          62784          -5.42%
BenchmarkChunkQueryable/foo/step_10m0s/Bigchunk/iterators/-8               62125          58528          -5.79%
BenchmarkChunkQueryable/foo/step_10m0s/Bigchunk/batches/-8                 65444          57063          -12.81%

benchmark                                                                  old bytes     new bytes     delta
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Bigchunk/matrixes/-8       84536361      84191244      -0.41%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Bigchunk/iterators/-8      21925200      21579928      -1.57%
BenchmarkChunkQueryable/rate(foo[1m])/step_1m0s/Bigchunk/batches/-8        22023864      21655798      -1.67%
BenchmarkChunkQueryable/foo/step_1m0s/Bigchunk/matrixes/-8                 84495292      84150110      -0.41%
BenchmarkChunkQueryable/foo/step_1m0s/Bigchunk/iterators/-8                21884737      21539648      -1.58%
BenchmarkChunkQueryable/foo/step_1m0s/Bigchunk/batches/-8                  21983480      21615403      -1.67%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Bigchunk/matrixes/-8      65872333      65527155      -0.52%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Bigchunk/iterators/-8     3671717       3322504       -9.51%
BenchmarkChunkQueryable/rate(foo[1m])/step_10m0s/Bigchunk/batches/-8       3769733       3400636       -9.79%
BenchmarkChunkQueryable/foo/step_10m0s/Bigchunk/matrixes/-8                65832268      65487046      -0.52%
BenchmarkChunkQueryable/foo/step_10m0s/Bigchunk/iterators/-8               3637489       3292078       -9.50%
BenchmarkChunkQueryable/foo/step_10m0s/Bigchunk/batches/-8                 4126843       3367903       -18.39%
```